Remove div creating styling errors
